合并 P4c 收口后的 Vulkan RHI 后端（130 commits，T4.6.5/P4c 全阶段），
使 dev 获得 Metal+Vulkan 双后端能力。RHI 层功能对等已达成
（F1–F8 落地，F9/F10 书面豁免），详见
Docs/2026-08-19-rhi-vulkan-metal-parity-plan.md 与
Engine/Graphics/RHI/Platforms/Vulkan/README.md §10。

冲突解决（5 文件）：
- .gitignore：采用 vulkan 侧 contents-list 写法，保留 dev 的 WFC
  资产再包含规则与 vulkan 的 SPIR-V/ReferenceImages 再包含规则
- Engine/Common/EngineConfig.h：按 dev 的 565e96a 保持不跟踪
  （CMake configure_file 继续生成）
- Nanite/DepthHistoryManager.cpp：三路 shader 加载共存——Vulkan 走
  HZBCopy SPIR-V，Emscripten 走 WGSL，其余走 HZBGeneration.metal
- RenderPipeline/Modules/LineBatchRenderer.cpp：WASM 早退 stub 保留，
  非 WASM 下 Vulkan SPIR-V / Metal .metal 双路共存
- RenderPipeline/Modules/ForwardSceneRenderer.cpp（9 处）：include 双取；
  Vulkan compute lighting set + DeferredLighting compute dispatch 路径
  与 dev 的 WASM WGSL 路径 / Metal 图形路径共存；
  populateWasmVertexInput 与 applyGBufferVertexInput 两个 helper 并存；
  push constant offset 统一取 vulkan 侧 0（dfecb1a：Vulkan 要求 4 字节
  对齐，Metal 后端忽略 offset 参数，已核实 MetalCommandBuffer 实现）

合并内含修复：
- Components/Particle.{h,cpp}：vulkan 分支将 CMAKE_CXX_STANDARD 升至
  20，原 std::hash<particle_id>（u32 别名的 std 再特化）在新 libc++
  下报 "explicit specialization after instantiation"。改为程序自定义
  particle_id_hash 显式传入 unordered_map（vulkan 分支预存隐患，
  其本地 c++17 旧构建目录未暴露）
- Engine/Graphics/Vulkan/Shaders/ 大小写冲突消除：dev 跟踪大写
  Shaders/（33 文件），vulkan 分支跟踪小写 shaders/（256 文件，为前者
  完整超集），macOS 大小写不敏感 FS 上两套索引条目指向同一物理文件。
  移除大写侧陈旧条目（磁盘文件不变，内容以小写侧 vulkan 新版为准，
  DeferredLighting.frag part25→part37 含 IBL 接线）。运行时
  ShaderRegistry/loadSpv 均按小写路径加载，无引用破坏

构建验证（本机 Darwin/arm64，Xcode 26.5 SDK）：
- 默认配置（Metal）Engine 目标编译链接通过
- ENABLE_VULKAN=ON Engine 目标编译链接通过
- emcmake + ENABLE_WEBGPU=ON（对齐 wasm-build.yml）Engine 目标编译通过

后续待办：README §10 Outstanding 的 4b-T4 引擎集成收口；
WFC/Emscripten/Vulkan 全量测试回归；Windows CI 观察 round-3/4 修复后状态。

+++ EngineTest/UnitTests/RHI/Core/HashBenchmark.cpp
@@ -88,7 +88,8 @@ std::vector<BenchmarkRenderItemKey> GenerateTestData(size_t count) {
     std::mt19937 gen(rd());
     std::uniform_int_distribution<u64> dist64(1, 1000000);
     std::uniform_int_distribution<u32> dist32(1, 1000);
-    std::uniform_int_distribution<u8> dist8(1, 10);
+    // MSVC 的 uniform_int_distribution 不允许 8 位类型（static_assert 限制）
+    std::uniform_int_distribution<u16> dist8(1, 10);
     
     for (size_t i = 0; i < count; ++i) {
         BenchmarkRenderItemKey key;
